update test suit (#2329)
In this PR, test suit is updated to 1) use `generate_random_numpy_array`
for generating inputs of a few functions, 2) avoid unnecessary
conversion to numpy array where possible, 3) to modify the format.

---------


Update build_conda_pkg.yml

+++ dpnp/tests/helper.py
@@ -147,7 +147,9 @@ def get_float_complex_dtypes(no_float16=True, device=None):
 def get_abs_array(data, dtype=None):
     if numpy.issubdtype(dtype, numpy.unsignedinteger):
         data = numpy.abs(data)
-    return numpy.array(data, dtype=dtype)
+    # it is better to use astype with the default casting=unsafe
+    # otherwise, we need to skip test for cases where overflow occurs
+    return numpy.array(data).astype(dtype)
 
 
 def get_all_dtypes(
